fix: Lock validify dependency to version 1.0.12 in axum-valid
This commit addresses a compilation error caused by the validify crate update to version 1.1.1. By explicitly setting the validify dependency to version 1.0.12 in the Cargo.toml of the axum-valid project, we avoid the issue where the 'prost' attribute cannot be found in the newer version of validify. This serves as a temporary solution until the underlying issue with validify 1.1.1 is resolved.